refactor : refactor Util API
