--- backend/src/main/java/com/ssafy/util/dto/Statistics.java
@@ -1,15 +1,0 @@
-package com.ssafy.util.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.Setter;
-
-@Setter
-@Getter
-@AllArgsConstructor
-public class Statistics {
-  int countOfAttraction;
-  int countOfMember;
-  int countOfPost;
-
-}
