+++ backend/src/main/java/com/ssafy/util/model/dto/PagingInfo.java
@@ -1,4 +1,4 @@
-package com.ssafy.util.dto;
+package com.ssafy.util.model.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
