+++ backend/src/main/java/com/ssafy/util/model/dto/Statistics.java
@@ -1,0 +1,13 @@
+package com.ssafy.util.model.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class Statistics {
+  private int countOfAttraction;
+  private int countOfMember;
+  private int countOfPost;
+
+}
