+++ backend/src/main/java/com/ssafy/util/model/dto/request/PageRequestDto.java
@@ -1,4 +1,4 @@
-package com.ssafy.util.dto;
+package com.ssafy.util.model.dto.request;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
