+++ backend/src/main/java/com/ssafy/util/model/dto/response/PageResponseDto.java
@@ -1,4 +1,4 @@
-package com.ssafy.util.dto;
+package com.ssafy.util.model.dto.response;
 
 import lombok.Builder;
 import lombok.Getter;
